+++ src/emmy/mathbox/physics.cljs
@@ -1,4 +1,5 @@
 (ns emmy.mathbox.physics
+  "ODE and physics-aware plotting functions using `MathBox`."
   (:require [emmy.viewer.physics :as ph]
             [mathbox.primitives :as mb]))