+++ src/emmy/viewer/compile.clj
@@ -26,7 +26,10 @@
   (and (map? m)
        (not (v/param-f? m))))
 
-(defn vectorize [f]
+(defn vectorize
+  "Given a function `f` (parametrized or not) of a single non-vector argument,
+  returns a similar version that takes `[x]` instead of `x`."
+  [f]
   (if (v/param-f? f)
     (update f :f
             (fn [f]